+++ api/src/models/mod.rs
@@ -1,21 +1,7 @@
-mod budget;
 mod budget_v2;
 mod user;
-mod bank_transaction;
-mod budget_category;
-mod budget_item;
-mod budget_transaction;
-mod budget_item_periodicity;
-mod budget_transaction_type;
 mod month_begins_on;
 
-pub use budget::*;
 pub use budget_v2::*;
 pub use user::*;
-pub use bank_transaction::*;
-pub use budget_category::*;
-pub use budget_item::*;
-pub use budget_transaction::*;
-pub use budget_item_periodicity::*;
-pub use budget_transaction_type::*;
 pub use month_begins_on::*;
